+++ src/install/isolated_install/Symlinker.rs
@@ -111,7 +111,8 @@ impl Symlinker {
                                     #[cfg(not(windows))]
                                     let is_dir = if let Ok(st) = bun_sys::lstat(self.dest.slice_z())
                                     {
-                                        bun_sys::posix::s_isdir(st.st_mode)
+                                        // `mode_t` is `u16` on android, `u32` on linux/darwin.
+                                        bun_sys::posix::s_isdir(st.st_mode as u32)
                                     } else {
                                         false
                                     };
